// src/features/users/types/user.types.ts
export interface UserName {
  title: string;
  first: string;
  last: string;
}

export interface UserLocation {
  street: {
    number: number;
    name: string;
  };
  city: string;
  state: string;
  country: string;
  postcode: string | number;
  coordinates: {
    latitude: string;
    longitude: string;
  };
  timezone: {
    offset: string;
    description: string;
  };
}

export interface UserPicture {
  large: string;
  medium: string;
  thumbnail: string;
}

export interface UserLogin {
  uuid: string;
  username: string;
  password: string;
  salt: string;
  md5: string;
  sha1: string;
  sha256: string;
}

export interface UserDob {
  date: string;
  age: number;
}

export interface UserId {
  name: string;
  value: string | null;
}

export interface User {
  gender: string;
  name: UserName;
  location: UserLocation;
  email: string;
  login: UserLogin;
  dob: UserDob;
  registered: UserDob;
  phone: string;
  cell: string;
  id: UserId;
  picture: UserPicture;
  nat: string;
}

export interface ApiInfo {
  seed: string;
  results: number;
  page: number;
  version: string;
}

export interface UsersApiResponse {
  results: User[];
  info: ApiInfo;
}

export interface UsersQueryParams {
  results?: number;
  page?: number;
  seed?: string;
}